DR tests

+++ database/migrations/2018_06_20_104348_create_dr_worksheets_table.php
@@ -19,7 +19,7 @@ class CreateDrWorksheetsTable extends Migration
 
             // This is the exatype id
             $table->integer('plate_id')->unsigned()->nullable()->index();
-            $table->integer('extraction_worksheet_id')->nullable()->unsigned()->index();
+            $table->bigInteger('extraction_worksheet_id')->nullable()->unsigned()->index();
 
             // 1 is in process
             // 2 is tested, results uploaded
